kembalikan api seperti sebelumnya,catatan transaksi

+++ database/migrations/2023_02_19_010524_create_detail_transaksis_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignUuid('id_barang');
             $table->ForeignUuid('id_transaksi');
             $table->string('qty');
-            $table->decimal('harga_item');
+            $table->integer('harga_item');
             $table->timestamps();
         });
     }
